modularize the songkick code more
